# WIP optimize json parser