--- a/airframe-json/src/main/scala/wvlet/airframe/json/JSONSource.scala
+++ b/airframe-json/src/main/scala/wvlet/airframe/json/JSONSource.scala
@@ -16,6 +16,8 @@ package wvlet.airframe.json
 import java.nio.ByteBuffer
 import java.nio.charset.StandardCharsets
 
+import wvlet.log.LogSupport
+
 /**
   *
   */
@@ -24,7 +26,7 @@ object JSONSource {
   def fromString(s: String): JSONSource                 = fromBytes(s.getBytes(StandardCharsets.UTF_8))
   def fromBytes(b: Array[Byte]): JSONSource             = fromBytes(b, 0, b.length)
   def fromBytes(b: Array[Byte], offset: Int, size: Int) = new JSONSource(b, offset, size)
-  def fromByteBuffer(b: ByteBuffer) = {
+  def fromByteBuffer(b: ByteBuffer): JSONSource = {
     val a       = new Array[Byte](b.remaining())
     val current = b.position()
     b.get(a, 0, a.length)
@@ -33,14 +35,19 @@ object JSONSource {
   }
 }
 
-final class JSONSource(private[this] val b: Array[Byte], private[this] val offset: Int, private[this] val size: Int) {
+final class JSONSource(private[this] val b: Array[Byte], private[this] val offset: Int, private[this] val size: Int)
+    extends LogSupport {
   assert(offset >= 0, s"The offset must be >= 0: ${offset}")
   assert(size >= 0, s"The size must be >= 0: ${size}")
   assert(offset + size <= b.length, s"The offset + size must be <= ${b.length}: ${offset}+${size}")
+  private val sb              = new StringBuilderExt
   def length: Int             = size
-  def apply(index: Int): Byte = b(index + offset)
+  def apply(index: Int): Byte = if (offset > 0) b(offset + index) else b(index)
   def substring(start: Int, end: Int): String = {
-    new String(b, offset + start, end - start, StandardCharsets.UTF_8)
+    (offset + start until end).foreach { i =>
+      sb.append(apply(i))
+    }
+    sb.getAndReset()
   }
 }