## Optimize JSON parser to make it faster
  - remove using mkString
  - remove using Try instead of try-catch 
  - remove IndexedSeq instead of ArrayBuffer
  - remove continue/break flag for while-loop
  - scanNumber: append char to StringBuilder with scan at the same time
  - substring: using same StringBuilder instance


## Remaing Tasks
- [ ] Add more Test Code for qualities.

## Benchmark Results

```
2019-07-18 22:57:27.996+0900  info [JSONBenchmark] -twitter.json        total:5.473 sec. , count:   10, avg:0.547 sec. , min:0.243 sec. , median:0.312 sec. , max:2.553 sec. 
  -airframe             total:1.138 sec. , count:  100, avg:0.011 sec. , min:3.624 msec., median:5.311 msec., max:0.118 sec. 
  -airframe scan        total:0.541 sec. , count:  100, avg:5.411 msec., min:1.625 msec., median:2.239 msec., max:0.034 sec. 
  -circe                total:0.523 sec. , count:  100, avg:5.233 msec., min:1.454 msec., median:2.186 msec., max:0.164 sec. 
  -jawn                 total:0.361 sec. , count:  100, avg:3.605 msec., min:1.467 msec., median:1.862 msec., max:0.039 sec. 
  -json4s-jackson rk / Ttotal:0.634 sec. , count:  100, avg:6.339 msec., min:1.522 msec., median:1.901 msec., max:0.343 sec. 
  -json4s-native        total:1.496 sec. , count:  100, avg:0.015 sec. , min:7.102 msec., median:8.904 msec., max:0.169 sec. 
  -uJson                total:0.761 sec. , count:  100, avg:7.607 msec., min:3.041 msec., median:4.112 msec., max:0.094 sec.  
2019-07-18 22:57:28.682+0900  info [JSONBenchmark] -boolean array       total:0.657 sec. , count:   10, avg:0.066 sec. , min:0.018 sec. , median:0.026 sec. , max:0.292 sec. 
  -airframe             total:0.101 sec. , count:  100, avg:1.005 msec., min:0.281 msec., median:0.346 msec., max:0.011 sec. 
  -airframe scan        total:0.057 sec. , count:  100, avg:0.575 msec., min:0.172 msec., median:0.197 msec., max:0.015 sec. 
  -circe                total:0.132 sec. , count:  100, avg:1.316 msec., min:0.157 msec., median:0.249 msec., max:0.014 sec. 
  -jawn                 total:0.078 sec. , count:  100, avg:0.779 msec., min:0.157 msec., median:0.206 msec., max:4.156 msec.
  -json4s-jackson rk / Ttotal:0.040 sec. , count:  100, avg:0.402 msec., min:0.240 msec., median:0.313 msec., max:3.136 msec.
  -json4s-native        total:0.077 sec. , count:  100, avg:0.769 msec., min:0.421 msec., median:0.501 msec., max:4.901 msec.
  -uJson                total:0.165 sec. , count:  100, avg:1.652 msec., min:0.215 msec., median:0.393 msec., max:0.017 sec.  
2019-07-18 22:57:31.140+0900  info [JSONBenchmark] -string array        total:2.379 sec. , count:   10, avg:0.238 sec. , min:0.149 sec. , median:0.193 sec. , max:0.690 sec. 
  -airframe             total:0.644 sec. , count:  100, avg:6.438 msec., min:3.386 msec., median:4.750 msec., max:0.024 sec. 
  -airframe scan        total:0.128 sec. , count:  100, avg:1.285 msec., min:0.718 msec., median:0.913 msec., max:5.633 msec.
  -circe                total:0.209 sec. , count:  100, avg:2.088 msec., min:0.884 msec., median:1.207 msec., max:0.013 sec. 
  -jawn                 total:0.143 sec. , count:  100, avg:1.427 msec., min:0.887 msec., median:1.239 msec., max:7.754 msec.
  -json4s-jackson rk / Ttotal:0.226 sec. , count:  100, avg:2.263 msec., min:0.948 msec., median:1.309 msec., max:0.030 sec. 
  -json4s-native        total:0.632 sec. , count:  100, avg:6.321 msec., min:4.003 msec., median:5.643 msec., max:0.016 sec. 
  -uJson                total:0.391 sec. , count:  100, avg:3.906 msec., min:0.976 msec., median:1.484 msec., max:0.034 sec.  

```