@smdmts 2019-08-22
This Travis-ci error is weird, error of `wvlet.airframe.http.recorder.HttpRecorderTest`

Same module's tests got no problem
https://travis-ci.org/smdmts/airframe/builds/574942384