@xerial 2019-08-22
@smdmts Thanks. The CI test becomes unstable after upgrading to sbt-1.3.0-RC4 and using parallel tests. You don’t need to worry about the test failure for now.