@xerial 2019-08-22
I’ve disabled the turbo mode of sbt in the latest master https://github.com/wvlet/airframe/commit/dfd7a8677bc7d8d6b3ddf92a464855f2464f35a7#diff-fdc3abdfd754eeb24090dbd90aeec2ce